Generate Logger service
1. Generated Logger service from CLI
2. Added log() method to Logger service
3. Created "should log input to console" test

This test is carried out against an actual instance of the Logger service and not a stub. Becasue this is a unit test we can test against the actual instance and not a mock.

If we were testing a different unit that has Logger Service as a dependency we would mock the service to avoid pollution.